requirements.txt
asgiref==3.7.2
Django==5.0.1
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.3
python-decouple==3.8
sqlparse==0.4.4
